requirements.txt
python-dotenv==1.0.0
playwright==1.41.1
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
httpx==0.24.1
